adicionando documentação do swagger

+++ VideoLocadora/src/main/java/io/github/gabryel/videolocadora/service/movie/MovieService.java
@@ -5,13 +5,15 @@ import io.github.gabryel.videolocadora.dto.movie.MovieSaveDTO;
 import io.github.gabryel.videolocadora.entity.MovieEntity;
 import io.github.gabryel.videolocadora.mapper.movie.MovieMapper;
 import io.github.gabryel.videolocadora.repository.MovieRepository;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MovieService {
-    
+
     private final MovieRepository movieRepository;
     private final MovieMapper movieMapper;
 
@@ -24,26 +26,39 @@ public class MovieService {
         MovieEntity movieEntity = movieMapper.toEntity(dto);
         return movieMapper.toDetailDto(movieRepository.save(movieEntity));
     }
-    
+
     public List<MovieDetailDTO> findAll() {
         return movieRepository.findAll().stream()
-            .map(movieMapper::toDetailDto)
-            .toList();
+                .map(movieMapper::toDetailDto)
+                .toList();
     }
-    
+
     public MovieDetailDTO findById(Long id) {
         return movieRepository.findById(id)
-            .map(movieMapper::toDetailDto)
-            .orElseThrow();
+                .map(movieMapper::toDetailDto)
+                .orElseThrow();
     }
-    
+
     public MovieDetailDTO update(Long id, MovieSaveDTO dto) {
         MovieEntity movieEntity = movieRepository.findById(id).orElseThrow();
         movieMapper.toEntity(dto);
         return movieMapper.toDetailDto(movieRepository.save(movieEntity));
     }
-    
-    public void delete(Long id) {
-        movieRepository.deleteById(id);
+
+    public boolean delete(Long id) {
+        Optional<MovieEntity> movie = movieRepository.findById(id);
+        if (movie.isPresent()) {
+            movieRepository.deleteById(id);
+            return true;
+        }
+
+        return false;
     }
+
+    public List<MovieDetailDTO> findAllMovies() {
+        return movieRepository.findAll().stream()
+                .map(movieMapper::toDetailDto)
+                .toList();
+    }
+
 }
